Updated test suite to allow for simultaneous test runs for all data providers
