--- tests/Collections/SampleDataCollection.cs
@@ -1,6 +1,0 @@
-using MagiQueryTests.Fixtures;
-
-namespace MagiQueryTests.Collections;
-
-[CollectionDefinition("SampleDataCollection")]
-public class SampleDataCollection : ICollectionFixture<TestDataFixture>;
