+++ tests/Collections/SampleDataCollections.cs
@@ -1,0 +1,10 @@
+using MagiQueryTests.Fixtures.Implementations;
+
+namespace MagiQueryTests.Collections;
+
+
+
+
+
+
+
